+++ files/index.html
@@ -1,0 +1,19 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <h1>我是h1标签</h1>
+    <img src="https://tenfei02.cfp.cn/creative/vcg/800/version23/VCG41175510742.jpg" alt="" style="width: 50%;"> 
+    <button id=b1> 点我 </button>
+</body>
+</html>
+
+<script>
+
+
+</script>
