+++ model/model.go
@@ -1,0 +1,5 @@
+package model
+
+type ReqQrcode struct {
+	Url string `form:"url"`
+}
